# plugin/pkg/cache: smarter locking